refactor(benchmarks): split buildWorkloads registry by domain (#845)
Extract one register module per catalog section under benchmarks/workloads/
and keep buildWorkloads as a thin fixed-order concatenator. Smoke/full
catalog freezes and budget ids stay byte-stable; measured fn behavior is
unchanged.
