+++ benchmarks/budgets.json
@@ -1,5 +1,5 @@
 {
-  "$comment": "PROVISIONAL budgets (plan M3): measured on an Apple M1 Pro dev machine (darwin arm64, bun 1.3.6) on 2026-07-11 via `bun run bench:json`; budget = median * 1.5 headroom, rounded up to ~2 significant figures. Re-baseline on the CI runner when a remote exists (plan M3->M4). Canvas workloads measure JS command-generation cost against a stub 2d context (bun has no raster) — see workloads.ts. Keys must exactly match the workload ids emitted by bench-json.ts; check-budgets.ts asserts coverage in both directions.",
+  "$comment": "PROVISIONAL budgets (plan M3): measured on an Apple M1 Pro dev machine (darwin arm64, bun 1.3.6) on 2026-07-11 via `bun run bench:json`; budget = median * 1.5 headroom, rounded up to ~2 significant figures. Re-baseline on the CI runner when a remote exists (plan M3->M4). Canvas workloads measure JS command-generation cost against a stub 2d context (bun has no raster) — see workloads/. Keys must exactly match the workload ids emitted by bench-json.ts; check-budgets.ts asserts coverage in both directions.",
   "budgets": {
     "pipeline scatter 1k": { "budgetMs": 2.9 },
     "svg render scatter 1k": { "budgetMs": 3.4 },
