--- benchmarks/workloads.ts
@@ -1,477 +1,0 @@
-/// <reference lib="dom" />
-/**
- * Shared workload table for the benchmark suite. ONE source of truth for
- * three consumers:
- *
- * - `pipeline.bench.ts` (mitata full suite + BENCH_SMOKE) — groups/benches
- *   are rebuilt from this table with the exact pre-M3 labels.
- * - `bench-json.ts` (M3) — runs the full-size workloads with a bounded
- *   deterministic-ish loop and emits `bench-results.json` (repo root) in
- *   github-action-benchmark customSmallerIsBetter format.
- * - `check-budgets.ts` (M3) — gates those results against
- *   `budgets.json`; `Workload.id` is the stable identifier both files key on.
- *
- * Chart fixtures live in `workload-specs.ts`. Data is generated from a
- * seeded PRNG so runs are comparable.
- */
-import {
-  buildCandidateStore,
-  FONT_METRICS,
-  MetricsTableMeasurer,
-  planStrata,
-  planTemporalAxis,
-  renderToSVGString,
-  runPipeline,
-} from "@ggsvelte/core";
-import { drawStratum } from "@ggsvelte/core/dom";
-import { MAX_BINNED_BREAKS } from "@ggsvelte/spec";
-
-import {
-  temporalLineSpec,
-  temporalFreeFacetSpec,
-  scatterSpec,
-  transformedScatterSpec,
-  transformedStatsSpec,
-  transformedFacetSpec,
-  coordPointSpec,
-  coordFixedSpec,
-  coordTessellationSpec,
-  maxBoundaryBinnedSpec,
-  nonPositionColorSpec,
-  responsiveGuideSpec,
-  mappedStyleSpec,
-  facetedBarsSpec,
-  stackedBarSpec,
-  lineSeriesSpec,
-  histogramSpec,
-  loessSpec,
-  densitySpec,
-} from "./workload-specs";
-
-export interface Workload {
-  /** Stable identifier for bench-results.json / budgets.json (M3 gating). */
-  id: string;
-  /** mitata group label (kept byte-identical to the pre-M3 suite). */
-  group: string;
-  /** mitata bench label (kept byte-identical to the pre-M3 suite). */
-  bench: string;
-  fn: () => unknown;
-}
-
-/**
- * Stub CanvasRenderingContext2D: measures the CANVAS COMMAND-GENERATION cost
- * of the batch renderers (JS-side), not rasterization — bun has no canvas.
- * Real paint cost is browser-side (component tests exercise it).
- */
-function stubContext(): CanvasRenderingContext2D {
-  const noop = (): void => {
-    /* raster cost lives in the browser; this stub measures JS command cost */
-  };
-  return {
-    fillStyle: "",
-    strokeStyle: "",
-    lineWidth: 1,
-    lineJoin: "round",
-    lineCap: "round",
-    globalAlpha: 1,
-    save: noop,
-    restore: noop,
-    beginPath: noop,
-    closePath: noop,
-    rect: noop,
-    clip: noop,
-    translate: noop,
-    clearRect: noop,
-    moveTo: noop,
-    lineTo: noop,
-    arc: noop,
-    fill: noop,
-    stroke: noop,
-    fillRect: noop,
-    strokeRect: noop,
-    setTransform: noop,
-  } as unknown as CanvasRenderingContext2D;
-}
-
-/** "1000" -> "1k"; sub-1k counts (smoke loess) stay plain. */
-function fmtK(n: number): string {
-  return n >= 1000 ? `${n / 1000}k` : `${n}`;
-}
-
-const opts = { width: 800, height: 500, maxMarks: 200_000 };
-
-/**
- * Build the workload list. `smoke` shrinks sizes (the CI bench-smoke job);
- * the JSON/budget path always uses `smoke = false` so ids are stable.
- * Order and labels mirror the pre-M3 mitata suite exactly.
- */
-export function buildWorkloads(smoke: boolean): Workload[] {
-  const workloads: Workload[] = [];
-  const sizes = smoke ? [1_000] : [1_000, 10_000, 100_000];
-
-  for (const n of sizes) {
-    const spec = scatterSpec(n);
-    const label = `${fmtK(n)} points`;
-    workloads.push(
-      {
-        id: `pipeline scatter ${fmtK(n)}`,
-        group: label,
-        bench: `runPipeline ${label}`,
-        fn: () => runPipeline(spec, opts),
-      },
-      {
-        id: `svg render scatter ${fmtK(n)}`,
-        group: label,
-        bench: `renderToSVGString ${label}`,
-        fn: () => renderToSVGString(spec, opts),
-      },
-    );
-  }
-
-  {
-    const n = smoke ? 1_000 : 100_000;
-    const spec = temporalLineSpec(n);
-    workloads.push({
-      id: `pipeline temporal-line ${fmtK(n)}`,
-      group: `temporal line ${fmtK(n)}`,
-      bench: `runPipeline temporal line ${fmtK(n)}`,
-      fn: () => runPipeline(spec, opts),
-    });
-  }
-
-  {
-    const measurer = new MetricsTableMeasurer(FONT_METRICS);
-    const input = {
-      aesthetic: "x" as const,
-      panelIndex: 0,
-      domain: [Date.UTC(1800, 0, 1), Date.UTC(2100, 0, 1)] as const,
-      kind: "date" as const,
-      orient: "horizontal" as const,
-      extentPx: 800,
-      reverse: false,
-      measurer,
-      fontSize: 11.5,
-      marginCapPx: 92,
-      config: {},
-    };
-    const resizeInput = {
-      ...input,
-      domain: [Date.UTC(1835, 0, 1), Date.UTC(2025, 0, 1)] as const,
-    };
-    workloads.push(
-      {
-        id: "temporal guide candidate-selection 300y",
-        group: "temporal guide candidate selection",
-        bench: "planTemporalAxis 300-year domain",
-        fn: () => planTemporalAxis(input),
-      },
-      {
-        id: "temporal guide resize-churn 191y",
-        group: "temporal guide responsive planning",
-        bench: "planTemporalAxis 191-year resize sequence",
-        fn: () => {
-          let previousInterval: string | null | undefined;
-          let result;
-          for (const extentPx of [320, 640, 1_200, 640, 320]) {
-            result = planTemporalAxis({
-              ...resizeInput,
-              extentPx,
-              ...(previousInterval !== undefined && { previousInterval }),
-            });
-            previousInterval = result.interval;
-          }
-          return result;
-        },
-      },
-      {
-        id: "temporal guide DST-heavy 3y",
-        group: "temporal guide zoned calendar planning",
-        bench: "planTemporalAxis DST-heavy datetime domain",
-        fn: () =>
-          planTemporalAxis({
-            ...input,
-            domain: [Date.UTC(2022, 0, 1), Date.UTC(2025, 0, 1)],
-            kind: "datetime",
-            config: { timezone: "America/New_York", dateBreaks: "1 month" },
-          }),
-      },
-    );
-  }
-
-  {
-    const spec = temporalFreeFacetSpec(smoke ? 10 : 100);
-    workloads.push({
-      id: `pipeline temporal free-facets ${smoke ? "10" : "100"}`,
-      group: `temporal free facets ${smoke ? "10" : "100"}`,
-      bench: `runPipeline temporal free facets ${smoke ? "10" : "100"}`,
-      fn: () => runPipeline(spec, { ...opts, width: 1_200, height: 900 }),
-    });
-  }
-
-  {
-    const spec = stackedBarSpec();
-    workloads.push(
-      {
-        id: "pipeline stacked-bars 50x4",
-        group: "stacked bars 50x4",
-        bench: "runPipeline stacked bars 50x4",
-        fn: () => runPipeline(spec, opts),
-      },
-      {
-        id: "svg render stacked-bars 50x4",
-        group: "stacked bars 50x4",
-        bench: "renderToSVGString stacked bars 50x4",
-        fn: () => renderToSVGString(spec, opts),
-      },
-    );
-  }
-
-  {
-    const perSeries = smoke ? 1_000 : 10_000;
-    const spec = lineSeriesSpec(perSeries);
-    const label = `line series 10x${fmtK(perSeries)}`;
-    workloads.push(
-      {
-        id: `pipeline line-series 10x${fmtK(perSeries)}`,
-        group: label,
-        bench: `runPipeline ${label}`,
-        fn: () => runPipeline(spec, opts),
-      },
-      {
-        id: `svg render line-series 10x${fmtK(perSeries)}`,
-        group: label,
-        bench: `renderToSVGString ${label}`,
-        fn: () => renderToSVGString(spec, opts),
-      },
-    );
-  }
-
-  // --- M2 facet + canvas workloads ------------------------------------------
-  {
-    const spec = facetedBarsSpec();
-    workloads.push(
-      {
-        id: "pipeline faceted-bars 50 panels",
-        group: "faceted bars 50 panels",
-        bench: "runPipeline faceted bars 50 panels",
-        fn: () => runPipeline(spec, opts),
-      },
-      {
-        id: "svg render faceted-bars 50 panels",
-        group: "faceted bars 50 panels",
-        bench: "renderToSVGString faceted bars 50 panels",
-        fn: () => renderToSVGString(spec, opts),
-      },
-    );
-  }
-
-  {
-    const n = smoke ? 1_000 : 100_000;
-    const spec = scatterSpec(n, "canvas");
-    const model = runPipeline(spec, opts);
-    const strata = planStrata(model.scene, model.layerBackends);
-    const canvasStratum = strata.find((s) => s.backend === "canvas");
-    if (canvasStratum === undefined) throw new Error("expected a canvas stratum");
-    const canvasBatches = canvasStratum.batches;
-    const ctx = stubContext();
-    const resolve = (c: string) => c;
-    const label = fmtK(n);
-    const groupLabel = `canvas scatter ${label} (stub ctx: JS command cost, no raster)`;
-    workloads.push(
-      {
-        id: `canvas cold scatter ${label}`,
-        group: groupLabel,
-        bench: `canvas cold ${label} (pipeline + plan + draw + candidate index)`,
-        fn: () => {
-          const m = runPipeline(spec, opts);
-          const plan = planStrata(m.scene, m.layerBackends);
-          for (const stratum of plan) {
-            if (stratum.backend === "canvas") drawStratum(ctx, m.scene, stratum.batches, resolve);
-          }
-          m.candidates.hitTest(400, 250);
-        },
-      },
-      {
-        id: `canvas redraw scatter ${label}`,
-        group: groupLabel,
-        bench: `canvas redraw ${label} (drawStratum only)`,
-        fn: () => {
-          drawStratum(ctx, model.scene, canvasBatches, resolve);
-        },
-      },
-      {
-        id: `hit-index build ${label}`,
-        group: groupLabel,
-        bench: `candidate index build ${label} points`,
-        fn: () => buildCandidateStore(model.scene).hitTest(400, 250),
-      },
-      {
-        id: `candidate lookup ${label}`,
-        group: groupLabel,
-        bench: `candidate nearest/group/rect ${label}`,
-        fn: () => {
-          const match = model.candidates.nearest(400, 250, { mode: "xy", maxDistance: 32 });
-          if (match !== null && match.xToken !== null) model.candidates.group(match.id, "x");
-          model.candidates.queryRect(200, 125, 600, 375);
-        },
-      },
-    );
-  }
-
-  // --- PR 3 pre-stat transform and binned-family workloads ------------------
-  {
-    const n = smoke ? 1_000 : 100_000;
-    for (const transform of ["identity", "log10", "sqrt"] as const) {
-      const spec = transformedScatterSpec(n, transform);
-      workloads.push({
-        id: `pipeline transform-${transform} ${fmtK(n)}`,
-        group: `position transform ${transform} ${fmtK(n)}`,
-        bench: `runPipeline transform ${transform} ${fmtK(n)}`,
-        fn: () => runPipeline(spec, opts),
-      });
-    }
-    for (const stat of ["smooth", "bin"] as const) {
-      const spec = transformedStatsSpec(n, stat);
-      workloads.push({
-        id: `pipeline transform-log10 ${stat} ${fmtK(n)}`,
-        group: `pre-stat log10 ${stat} ${fmtK(n)}`,
-        bench: `runPipeline log10 ${stat} ${fmtK(n)}`,
-        fn: () => runPipeline(spec, opts),
-      });
-    }
-    const facets = transformedFacetSpec(n, smoke ? 10 : 100);
-    workloads.push({
-      id: `pipeline transform-log10 facets-${smoke ? "10" : "100"} ${fmtK(n)}`,
-      group: `shared transform cache ${smoke ? "10" : "100"} facets ${fmtK(n)}`,
-      bench: `runPipeline log10 shared facets ${fmtK(n)}`,
-      fn: () => runPipeline(facets, { ...opts, width: 1_200, height: 900 }),
-    });
-    const binned = maxBoundaryBinnedSpec(n);
-    workloads.push({
-      id: `pipeline binned-${MAX_BINNED_BREAKS} ${fmtK(n)}`,
-      group: `binned ${MAX_BINNED_BREAKS} boundaries ${fmtK(n)}`,
-      bench: `runPipeline max-boundary binned ${fmtK(n)}`,
-      fn: () => runPipeline(binned, opts),
-    });
-  }
-
-  // --- PR 4 post-stat coordinate projection + bounded tessellation ---------
-  {
-    const n = smoke ? 1_000 : 100_000;
-    for (const transform of ["identity", "log10"] as const) {
-      const spec = coordPointSpec(n, transform);
-      workloads.push({
-        id: `pipeline coord-${transform} points ${fmtK(n)}`,
-        group: `post-stat coordinate ${transform} ${fmtK(n)}`,
-        bench: `runPipeline coord ${transform} points ${fmtK(n)}`,
-        fn: () => runPipeline(spec, opts),
-      });
-    }
-    const vertices = smoke ? 1_000 : 10_000;
-    const tessellated = coordTessellationSpec(vertices);
-    workloads.push({
-      id: `pipeline coord-tessellation ${fmtK(vertices)}`,
-      group: `post-stat coordinate tessellation ${fmtK(vertices)}`,
-      bench: `runPipeline worst-case coord tessellation ${fmtK(vertices)}`,
-      fn: () => runPipeline(tessellated, opts),
-    });
-  }
-
-  // --- PR 5 generic non-position scale families ----------------------------
-  {
-    const n = smoke ? 1_000 : 100_000;
-    for (const family of ["log10", "binned", "manual"] as const) {
-      const spec = nonPositionColorSpec(n, family);
-      workloads.push({
-        id: `pipeline color-${family} ${fmtK(n)}`,
-        group: `non-position color ${family} ${fmtK(n)}`,
-        bench: `runPipeline color ${family} ${fmtK(n)}`,
-        fn: () => runPipeline(spec, opts),
-      });
-    }
-  }
-
-  // --- PR 6 complete mapped style vectors ----------------------------------
-  {
-    const n = smoke ? 1_000 : 100_000;
-    const spec = mappedStyleSpec(n);
-    workloads.push({
-      id: `pipeline mapped-style ${fmtK(n)}`,
-      group: `mapped style vectors ${fmtK(n)}`,
-      bench: `runPipeline mapped style vectors ${fmtK(n)}`,
-      fn: () => runPipeline(spec, opts),
-    });
-  }
-
-  // --- PR 7 responsive guide replanning -----------------------------------
-  {
-    const n = smoke ? 1_000 : 10_000;
-    const spec = responsiveGuideSpec(n);
-    let narrow = false;
-    workloads.push({
-      id: `pipeline responsive-guides resize ${fmtK(n)}`,
-      group: `responsive guide resize ${fmtK(n)}`,
-      bench: `runPipeline responsive guide resize ${fmtK(n)}`,
-      fn: () => {
-        narrow = !narrow;
-        return runPipeline(spec, { ...opts, width: narrow ? 420 : 800 });
-      },
-    });
-  }
-
-  // --- PR 8 fixed-aspect allocation under responsive resize ----------------
-  {
-    const n = smoke ? 1_000 : 10_000;
-    const spec = coordFixedSpec(n);
-    let narrow = false;
-    workloads.push({
-      id: `pipeline coord-fixed resize ${fmtK(n)}`,
-      group: `fixed-aspect resize ${fmtK(n)}`,
-      bench: `runPipeline coord_fixed resize ${fmtK(n)}`,
-      fn: () => {
-        narrow = !narrow;
-        return runPipeline(spec, {
-          ...opts,
-          width: narrow ? 360 : 900,
-          height: narrow ? 640 : 420,
-        });
-      },
-    });
-  }
-
-  // --- M2 statistical workloads (plan: bin 100k, loess 5k, density 100k) ----
-  {
-    const n = smoke ? 1_000 : 100_000;
-    const spec = histogramSpec(n);
-    workloads.push({
-      id: `pipeline histogram ${fmtK(n)}`,
-      group: `histogram ${fmtK(n)}`,
-      bench: `runPipeline histogram ${fmtK(n)}`,
-      fn: () => runPipeline(spec, opts),
-    });
-  }
-
-  {
-    const n = smoke ? 500 : 5_000;
-    const spec = loessSpec(n);
-    workloads.push({
-      id: `pipeline loess ${fmtK(n)}`,
-      group: `loess smooth ${n} rows`,
-      bench: `runPipeline loess+se ${n} rows`,
-      fn: () => runPipeline(spec, opts),
-    });
-  }
-
-  {
-    const n = smoke ? 1_000 : 100_000;
-    const spec = densitySpec(n);
-    workloads.push({
-      id: `pipeline density ${fmtK(n)}`,
-      group: `density ${fmtK(n)}`,
-      bench: `runPipeline density ${fmtK(n)}`,
-      fn: () => runPipeline(spec, opts),
-    });
-  }
-
-  return workloads;
-}
